Created Toolbar Component
